Make ImmutableAttributes more efficient
Keep the mapping from attribute to value in a map
instead of going up the parent chain on every request.
This is faster since attributes are queried much more
often than they are created and creation is de-duplicated,
so there are only a few such attribute containers in the
whole build.